+++ modules/agerun_assignment_instruction_evaluator.c
@@ -139,7 +139,7 @@ static data_t* _evaluate_expression_ast(expression_evaluator_t *mut_expr_evaluat
 /**
  * Evaluates an assignment instruction AST node
  */
-bool ar__assignment_instruction_evaluator__evaluate(
+bool ar_assignment_instruction_evaluator__evaluate(
     expression_evaluator_t *mut_expr_evaluator,
     data_t *mut_memory,
     const instruction_ast_t *ref_ast