+++ modules/agerun_parse_instruction_evaluator.c
@@ -267,7 +267,7 @@ static data_t* _parse_value_string(const char *value_str) {
 /**
  * Evaluates a parse instruction
  */
-bool ar__parse_instruction_evaluator__evaluate(
+bool ar_parse_instruction_evaluator__evaluate(
     expression_evaluator_t *mut_expr_evaluator,
     data_t *mut_memory,
     const instruction_ast_t *ref_ast